Add server-specific .roo file loading.
There are now 3 OpenMeridian servers open and due to differing update
schedules and content, the same .roo file may be different on each
server. Currently this is the case for Raza, where the version of the
city (7 rooms anyway) is different on 103/112 and 105.

To allow the single Ogre client to work on all 3 servers, I have changed
.roo loading to first check a server-specific folder (currently this is
the server name/number) for any .roo files before checking the parent
rooms directory. 103 would have its changed rooms in \rooms\103\, 105 in
\rooms\105\ etc.

In BaseClient.Connect(), the ServerSubFolder field of ResourceManager
is set to the selected server's name (103, 105 etc.). Server-specific
room files (and later, other legacy resources) are then added to the
dictionaries.

Preloading legacy resource occurs after this point, instead of on
client startup.