Use epsilon check for testing floor/wall heights.
These heights can be derived from slope equations and end up very
slightly different (e.g. 1551.999... compared to 1552.0). This can lead
to rare cases where texture coordinates are calculated incorrectly and
as this behavior is present in the classic client, the fix was to adjust
texture offsets in the rooms when it occurred.

Removed the two offset fixes from marion.roo (inn walls).