Changed Options button inside login window.
Connect button is now on the left with a small button with a wrench icon
to the right. This implements the ability to use regular buttons with a
picture. Not the most flexible implementation, since the picture will
fill a rectangle with a distance of 5 pixels from the borders of the
button. It seems like the most practical way to do it though.