Added utility spells to the default action button layout.
Phase is on grave/Zirkumflex (the key to the left of 1), since its a
great compromise between out of the way and quick to reach. You don't
hit it by accident, but you can hit it quickly and accurately if you
need to.

Blink is on 0 (located next to grave/Zirkumflex on the action button
layout)

Meditate is on F (since it restores mana while R restores vigor)

Conveyance is on T (above the loot key G)

I expect the R and F keys to be occupied by something else rather soon,
but putting Rest and Meditate on such prominent keys might highlight
their importance to new players.