Replicate highlighting effect from classic client.
Server will send LIGHT_FLAG_HIGHLIGHT (value 0x08) in the light flags
for any object which should have a highlight effect applied to it (i.e.
a small circle of light underneath). This is currently applied to signs,
to draw attention to them and encourage newbies to click on them.

The classic client handles this flag by capping the max range the light
can have, and not attenuating the light effect on floor textures. The
light is sent with max intensity, the end result being a small circle of
relatively bright light.

The way I've attempted to replicate it in ogre is by setting the range
to 12% of normal - the end result looks the same as classic.