Enhanced looting functionality
Unless a lootable item is selected, the loot window will now open if the
loot button is pressed and close again upon a second press. If a
lootable item is selected, it will be looted without loot window toggle
instead. If the modifier key is held, everything will be looted without
toggle as before.