Add registered user/allowed user checking
Added a new list of users, 'AllowedUsers', which have bot chat but not
admin ability.

Removed the requirement for users to have OPs in the channel to use the
bot, instead users must be in the admin or allowed users list, and also
have a registered nickname. The bot keeps a list of the current users in
the channel and whether or not they are registered.

The bot uses the extended who message (WhoX) to determine whether or
not users are registered. A queue of messages is used to avoid flooding.