Add new 112/200 domain, increment Ogre client vers.
Servers 112 and 200 are moving to a new domain, changing from arantis.eu
to meridian59.de. Updated the connection profiles and also added a check
for the old host during connection profile load so that we can update
existing clients without moving 112/200 in their connection list.
Incremented the Ogre client version since we will need players to update
so they have the new domain.