Add 'Abandon guildhall' button to Guild UI.
Guild UI was missing an 'Abandon guildhall' button to drop the guild's
hall. Added this, and also added text which is displayed instead of the
guildhall options if the user has no hall.

If the user shouldn't have the tabs for GuildMaster and Shield, remove
them from being displayed. Replace them if the user then tries to access
the Guild menu as a guildmaster.