Add 'always on top' confirm popup, OK button.
The confirm popup dialog will now disallow interacting with the rest of
the UI until one of the buttons is selected. Added a centered OK button
and an option to choose either 'Yes/No' or just 'OK' for the confirm
dialog button options. The 'OK' button would be used for e.g. error
message that just requires user to acknowledge it (incorrect PW when
logging in).