Fix irregular right-harnd border on patcher UI.
Using FormBorderStyles FixedSingle or Fixed3D causes the right-hand
border of the UI to stick out a few pixels past the title bar. Using
Sizeable instead (the default I guess since it didn't add a property for
it in designer) fixes this issue. Added a minimum size so the UI can't
actually be resized.

The other changes were made by VS when i changed the border, seems
that's the way it wants it laid out.