Switch from MD5 to SHA256, hashing/end transition pause.
Switch from MD5 to SHA256 for hashing files as the .NET MD5 algorithm
can cause an exception if user has FIPS enabled. Added a try-catch
to worker creation to handle cases where an outdated Win XP system may
not be able to create the SHA256 hasher.Move Worker creation+start to
Main to handl the exception early, before downloading patch info.

Added a short wait time transitioning from hashing to downloading, and
a wait time before closing the updater. Removed the messagebox with the
'client updated' info, that is now displayed in the status label.

Compressed the splash image for the picturebox, halves exe size.