Add quest minimap changes.
Add yellow rings for NPCs with quests to give user, green rings for NPCs
who are the destination of a user's current quest, and change red dot
for monsters to a bright purple dot if mob is target of a current kill
quest. Mob dot looks similar to existing dot for other user's minions,
but is brighter and mob will additionally have the purple ! above it if
part of a quest.