Reduce slow calls in bitmap processing loop.
The call to Color.ToArgb() for some reason is exceptionally slow, which
is compounded by its presense in the inner loop of the bitmap->pixel
code. Obtaining these values and storing them in an array leads to a
large speedup when processing bitmaps (e.g. loading images for use in
BGF editor).