Resize MultiLineEditBoxes to fit text where possible.
Object descriptions (mainly for items) and spell descriptions tend to
force a scrollbar on the inspection window as the window isn't large
enough to display all the text. At the same time, a default larger
window would look strange for short descriptions. This change
dynamically resizes the window so more of the text will fit, from the
current min up to the current max size.

Object inspection windows with inscriptions aren't resized as this is
less of a problem, and the main focus (inscription editbox) on the most
common items with inscriptions already uses the available max space.