Add 'export storyboard' feature to BGF editor.
Exports a 'storyboard' of a BGF. A storyboard is a single bmp containing
all the frames of the BGF, arranged vertically in order of frame number.
Useful for making graphical changes to all frames of a bgf. Exporting
the storyboard also outputs an xml file with the filenames of the frames
omitted, and extra fields for total height and width of the storyboard
bmp.