Allow users to slightly increase client brightness.
Adds an slider to the Options UI under the Engine tab allowing the user
to alter the ambient + directional brightness. The lowest setting is the
current brightness, and max is 1.8x the current brightness. On the
highest setting, Ogre displays rooms more closely to how classic does
(though classic tends to be too bright).