Fix UI windows visibility setting on close.
If the user quits the game using 'quit' and then closes the client (e.g.
with alt-f4) the window positions and visibility will be saved when
OgreClient::Cleanup() is called. Because the client is not in game mode,
all UI visibility will be saved as not visible. The fix is to check if
the game is in UIMode::Playing prior to saving setting visibility in
config for saving.