Show "Phased" notification when phased.
- Show "Phased" rather than "Paralyzed" when phased out. Works by adding
a listener on AvatarBuffs and taking action when the phase buff is added
and removed.
- Phase buff being active prevents "Paralyzed" from being shown.
Removing phase does a check for Paralyze being active to handle the hold
effect given when phase is cast again.
- Clear effects on Reset() in DataController.cs (i.e. when Quit is
used). This was previously done in Invalidate() which caused issues
removing effects that are not sent again during the server's system
save, but Reset() relied on it to reset effects when quitting to char
select. Fixes the edge case where quitting to char select while phased
would leave the whiteout effect and paralysis when logging back in.